Move recording functionality to new <AudioRecorder> component
It makes sense for the audio recording functionality to be separated from the <WebMidi> component, although it may still be preferable for the recording controls to continue to share a right-hand menu tab with the MIDI settings -- the tab selector is getting a bit crowded.

The two components were initially combined because the audio recording code was derived from a draft of a MIDI recording feature, which essentially eavesdropped on the Web MIDI message generation functionality (but wasn't limited only to browsers with Web MIDI support). Audio recording uses a completely different mechanism from Web MIDI, of course.

Relatedly, there seems to be no need to introduce a functionality check for the audio recording feature, because unlike the Web MIDI API, the MediaStream Recording API is supported on all browsers on which the Pianolatron might be reasonably expected to run.