Enable tooltipAction on panel switcher buttons
This necessitated some changes to the styling and implementation of the panel switcher buttons, the most notable of which was making them IconButton components with custom styles. There's still a bit of a UX issue with the panel switcher buttons not having visible text labels (ironically this now more of an issue for visitors who aren't using screen readers); in particular it makes it awkward to specify which button to press when writing documentation. But at least the active tooltips are more immediately and visibly displayed when moused over than passive tooltips. Are such active tooltips also necessary for the transport control buttons? Maybe not, if only because their iconography is more conventional and readily understood.